GITHUB-APD-165: reafcto the way of creating markdown-it to avoid to repeat ourself
